Collapse Megatron shared API into train module
